Add existing file

+++ ai_12/mariia_stepova/Epic2/vns_lab_1_task_1_double_mariia_stepova.cpp
@@ -3,9 +3,9 @@
 
 int main()
 {
-    long double a = 1000;
-    long double b = 0.0001;
-    long double x = pow(a+b,2) , y = pow(a, 2) + 2*a*b , z = pow(b, 2) , R;
+    float a = 1000;
+    float b = 0.0001;
+    float x = pow(a+b,2) , y = pow(a, 2) + 2*a*b , z = pow(b, 2) , R;
 
     R = (x-y)/z;
